power: ti: bq2597x_charger: Fix floating-point type conversion
../drivers/power/supply/ti/bq2597x_charger.c: In function 'bq2597x_get_adc_data':
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
 1116 |                         t = t * (1 + 1.803 * 0.001);
      |                         ~~^~~~~~~~~~~~~~~~~~~~~~~~~
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types
../drivers/power/supply/ti/bq2597x_charger.c:1116:27: error: '-mgeneral-regs-only' is incompatible with the use of floating-point types

+++ drivers/power/supply/ti/bq2597x_charger.c
@@ -1113,7 +1113,7 @@ static int bq2597x_get_adc_data(struct bq2597x *bq, int channel,  int *result)
 		t = val_l + (val_h << 8);
 		/* vbat need calibration read by NU2105 */
 		if (channel == ADC_VBAT)
-			t = t * (1 + 1.803 * 0.001);
+			t = t * (int)(1.000 + 1.803 * 0.001);
 		*result = t;
 	} else {
 		ret = bq2597x_read_word(bq, ADC_REG_BASE + (channel << 1), &val);
